// data/generated-stars.ts
export const generatedStars: Record<string, {
  stars: { name: string; top: string; left: string; size: number }[];
  lines: [number, number][];
}> = {
// ...
  "aquila": { stars: [
      { name: "Altair", top: "30.4%", left: "33.6%", size: 8 },
      { name: "Tarazed", top: "23.3%", left: "38.8%", size: 5 },
      { name: "Okab", top: "10.0%", left: "85.7%", size: 5 },
      { name: "65The Aql", top: "69.9%", left: "10.0%", size: 5 },
      { name: "30Del Aql", top: "53.9%", left: "62.6%", size: 5 },
      { name: "16Lam Aql", top: "86.5%", left: "84.8%", size: 5 },
      { name: "Alshain", top: "40.4%", left: "28.4%", size: 4 },
      { name: "55Eta Aql", top: "62.5%", left: "31.6%", size: 4 },
      { name: "12    Aql", top: "90.0%", left: "90.0%", size: 4 }
  ], lines: [[6,0], [0,1], [1,4], [4,2], [2,4], [4,7], [7,3], [3,7], [7,4], [4,5], [5,8]] },
// ...
  "cassiopeia": { stars: [
      { name: "Cih", top: "43.1%", left: "53.9%", size: 7 },
      { name: "Schedar", top: "90.0%", left: "66.2%", size: 7 },
      { name: "Caph", top: "60.7%", left: "90.0%", size: 7 },
      { name: "Ruchbah", top: "48.5%", left: "31.7%", size: 5 },
      { name: "Segin", top: "10.0%", left: "10.0%", size: 5 }
  ], lines: [[4,3], [3,0], [0,1], [1,2]] },
// ...
  "cygnus": { stars: [
      { name: "Deneb", top: "31.7%", left: "34.4%", size: 8 },
      { name: "Sadr", top: "48.6%", left: "49.3%", size: 7 },
      { name: "Aljanah", top: "69.8%", left: "30.7%", size: 7 },
      { name: "Fawaris", top: "32.2%", left: "78.2%", size: 5 },
      { name: "Albireo", top: "90.0%", left: "89.2%", size: 5 },
      { name: "64Zet Cyg", top: "82.4%", left: "10.0%", size: 5 },
      { name: "10Iot2Cyg", top: "10.0%", left: "90.0%", size: 4 },
      { name: "21Eta Cyg", top: "66.0%", left: "69.4%", size: 4 }
  ], lines: [[4,7], [7,1], [1,2], [2,5], [5,2], [2,1], [1,0], [0,1], [1,3], [3,6]] },
// ...
  "gemini": { stars: [
      { name: "Pollux", top: "32.5%", left: "10.0%", size: 8 },
      { name: "Alhena", top: "76.7%", left: "63.4%", size: 7 },
      { name: "Castor B", top: "17.9%", left: "18.5%", size: 5 },
      { name: "Tejat", top: "53.5%", left: "75.1%", size: 5 },
      { name: "Mebsuta", top: "43.5%", left: "58.5%", size: 5 },
      { name: "Propus", top: "53.5%", left: "81.5%", size: 5 },
      { name: "Alzirr", top: "90.0%", left: "57.5%", size: 5 },
      { name: "Wasat", top: "55.5%", left: "29.9%", size: 4 },
      { name: "77Kap Gem", top: "46.3%", left: "10.7%", size: 4 },
      { name: "54Lam Gem", top: "76.2%", left: "31.5%", size: 4 },
      { name: "34The Gem", top: "10.0%", left: "51.5%", size: 4 },
      { name: "60Iot Gem", top: "33.4%", left: "25.5%", size: 4 },
      { name: "Mekbuda", top: "60.9%", left: "42.6%", size: 4 },
      { name: "69Ups Gem", top: "36.8%", left: "17.4%", size: 4 },
      { name: "18Nu  Gem", top: "62.2%", left: "70.4%", size: 4 },
      { name: "1    Gem", top: "50.6%", left: "90.0%", size: 4 },
      { name: "62Rho Gem", top: "18.3%", left: "22.8%", size: 4 },
      { name: "46Tau Gem", top: "24.1%", left: "37.0%", size: 4 }
  ], lines: [[15,5], [5,3], [3,4], [4,14], [14,4], [4,17], [17,16], [16,2], [2,16], [16,17], [17,10], [10,17], [17,11], [11,13], [13,0], [0,13], [13,8], [8,13], [13,7], [7,12], [12,1], [1,12], [12,7], [7,9], [9,6]] },
// ...
  "leo": { stars: [
      { name: "Regulus", top: "66.2%", left: "73.0%", size: 8 },
      { name: "Denebola", top: "55.8%", left: "10.0%", size: 7 },
      { name: "Zosma", top: "32.0%", left: "31.9%", size: 5 },
      { name: "Ras Elased Australis", top: "18.9%", left: "87.1%", size: 5 },
      { name: "Chertan", top: "52.4%", left: "31.8%", size: 5 },
      { name: "Adhafera", top: "20.4%", left: "67.8%", size: 5 },
      { name: "30Eta Leo", top: "47.0%", left: "73.6%", size: 5 },
      { name: "Subra", top: "74.5%", left: "90.0%", size: 4 },
      { name: "Rasalas", top: "10.0%", left: "82.7%", size: 4 },
      { name: "78Iot Leo", top: "72.0%", left: "25.7%", size: 4 },
      { name: "77Sig Leo", top: "90.0%", left: "27.5%", size: 4 }
  ], lines: [[3,8], [8,5], [5,6], [6,0], [0,7], [7,0], [0,4], [4,9], [9,10], [10,9], [9,4], [4,1], [1,2]] },
// ...
  "lyra": { stars: [
      { name: "Vega", top: "10.0%", left: "90.0%", size: 8 },
      { name: "Sulafat", top: "90.0%", left: "10.0%", size: 5 },
      { name: "Sheliak", top: "81.2%", left: "42.2%", size: 4 },
      { name: "12Del2Lyr", top: "34.7%", left: "26.1%", size: 4 },
      { name: "6Zet1Lyr", top: "25.5%", left: "61.5%", size: 4 }
  ], lines: [[0,4], [4,2], [2,1], [1,3], [3,4]] },
// ...
  "orion": { stars: [
      { name: "Rigel", top: "86.1%", left: "65.9%", size: 8 },
      { name: "Betelgeuse", top: "44.4%", left: "26.3%", size: 8 },
      { name: "Bellatrix", top: "47.2%", left: "55.6%", size: 7 },
      { name: "Alnilam", top: "67.4%", left: "44.8%", size: 7 },
      { name: "Alnitak", top: "69.4%", left: "40.4%", size: 7 },
      { name: "Saiph", top: "90.0%", left: "33.6%", size: 7 },
      { name: "Mintaka", top: "65.0%", left: "48.9%", size: 7 },
      { name: "Tabit", top: "45.6%", left: "90.0%", size: 5 },
      { name: "Meissa", top: "37.6%", left: "45.9%", size: 5 },
      { name: "3Pi 4Ori", top: "49.2%", left: "88.7%", size: 4 },
      { name: "8Pi 5Ori", top: "57.6%", left: "85.7%", size: 4 },
      { name: "9Omi2Ori", top: "28.1%", left: "83.6%", size: 4 },
      { name: "61Mu  Ori", top: "38.4%", left: "19.3%", size: 4 },
      { name: "2Pi 2Ori", top: "40.4%", left: "89.2%", size: 4 },
      { name: "54Chi1Ori", top: "10.0%", left: "27.1%", size: 4 },
      { name: "67Nu  Ori", top: "24.7%", left: "14.3%", size: 4 },
      { name: "70Xi  Ori", top: "26.2%", left: "10.0%", size: 4 },
      { name: "10Pi 6Ori", top: "59.6%", left: "81.5%", size: 4 },
      { name: "62Chi2Ori", top: "10.4%", left: "17.8%", size: 3 },
      { name: "11    Ori", top: "23.0%", left: "75.6%", size: 3 },
      { name: "15    Ori", top: "22.5%", left: "70.6%", size: 3 }
  ], lines: [[4,5], [5,0], [0,6], [6,2], [2,8], [8,1], [1,12], [12,16], [16,15], [15,14], [14,18], [18,16], [16,12], [12,1], [1,4], [4,3], [3,6], [6,2], [2,7], [7,9], [9,10], [10,17], [17,10], [10,9], [9,7], [7,13], [13,11], [11,19], [19,20]] },
// ...
  "scorpius": { stars: [
      { name: "Antares", top: "32.6%", left: "67.5%", size: 8 },
      { name: "Shaula", top: "69.1%", left: "20.3%", size: 7 },
      { name: "Sargas", top: "89.2%", left: "17.6%", size: 7 },
      { name: "Dschubba", top: "19.6%", left: "88.9%", size: 7 },
      { name: "Larawag", top: "59.5%", left: "52.2%", size: 7 },
      { name: "Kap Sco", top: "75.6%", left: "13.7%", size: 7 },
      { name: "Lesath", top: "69.7%", left: "22.4%", size: 5 },
      { name: "Paikauhale", top: "38.7%", left: "62.8%", size: 5 },
      { name: "Fang", top: "31.5%", left: "90.0%", size: 5 },
      { name: "Alniyat", top: "29.8%", left: "73.6%", size: 5 },
      { name: "Iot1Sco", top: "79.4%", left: "10.0%", size: 5 },
      { name: "Xamidimura", top: "72.3%", left: "51.0%", size: 5 },
      { name: "Eta Sco", top: "90.0%", left: "36.1%", size: 5 },
      { name: "Zet2Sco", top: "87.0%", left: "49.0%", size: 4 },
      { name: "Acrab B", top: "10.0%", left: "85.2%", size: 3 }
  ], lines: [[14,9], [9,3], [3,9], [9,8], [8,9], [9,0], [0,7], [7,4], [4,11], [11,13], [13,12], [12,2], [2,10], [10,5], [5,1], [1,6]] },
// ...
  "ursa-major": { stars: [
      { name: "Alioth", top: "28.0%", left: "23.5%", size: 7 },
      { name: "Dubhe", top: "13.3%", left: "51.3%", size: 7 },
      { name: "Alkaid", top: "44.9%", left: "10.0%", size: 7 },
      { name: "Mizar", top: "30.6%", left: "16.0%", size: 7 },
      { name: "Merak", top: "26.9%", left: "51.8%", size: 7 },
      { name: "Phecda", top: "33.8%", left: "38.7%", size: 7 },
      { name: "52Psi UMa", top: "57.1%", left: "49.8%", size: 5 },
      { name: "Tania Australis", top: "64.7%", left: "61.7%", size: 5 },
      { name: "Talitha", top: "48.1%", left: "82.7%", size: 5 },
      { name: "25The UMa", top: "38.9%", left: "74.2%", size: 5 },
      { name: "Megrez", top: "25.3%", left: "33.2%", size: 5 },
      { name: "Muscida", top: "15.9%", left: "90.0%", size: 5 },
      { name: "Tania Borealis", top: "61.1%", left: "63.1%", size: 5 },
      { name: "Alula Borealis", top: "86.0%", left: "47.6%", size: 5 },
      { name: "Alkaphrah", top: "50.4%", left: "81.6%", size: 4 },
      { name: "23    UMa", top: "10.0%", left: "74.6%", size: 4 },
      { name: "Taiyangshou", top: "48.8%", left: "40.6%", size: 4 },
      { name: "29Ups UMa", top: "20.2%", left: "69.6%", size: 4 },
      { name: "Alula Australis", top: "90.0%", left: "47.7%", size: 4 }
  ], lines: [[2,3], [3,0], [0,10], [10,5], [5,16], [16,13], [13,18], [18,13], [13,16], [16,6], [6,7], [7,12], [12,7], [7,6], [6,16], [16,5], [5,4], [4,1], [1,10], [10,1], [1,15], [15,11], [11,17], [17,4], [4,17], [17,9], [9,14], [14,8]] },

  "ursa-minor": { stars: [
      { name: "Polaris", top: "10.0%", left: "10.0%", size: 7 },
      { name: "Kochab", top: "79.3%", left: "90.0%", size: 7 },
      { name: "Pherkad", top: "90.0%", left: "86.6%", size: 5 },
      { name: "22Eps UMi", top: "43.2%", left: "76.8%", size: 4 },
      { name: "16Zet UMi", top: "62.6%", left: "83.9%", size: 4 },
      { name: "Yildun", top: "22.3%", left: "71.6%", size: 4 },
      { name: "21Eta UMi", top: "72.0%", left: "80.1%", size: 3 }
  ], lines: [[0,5], [5,3], [3,4], [4,1], [1,2], [2,6], [6,4]] },
// ...
};